Revert "Add message list to campaigns" (#276)
Revert "Add message list to campaigns (#271)"

This reverts commit e877b82c83b75cd34d1f96b5bf27d2369d000965.
