+++ lib/bike_brigade_web/live/sms_message_live/conversation_component.html.heex
@@ -2,6 +2,7 @@
   id={"conversation-component:#{@rider.id}"}
   class={[
     "sm:flex-1 px-1 flex-col bg-white shadow sm:rounded-lg sm:flex",
+    "max-h-[calc(100vh-6rem)] sm:max-h-[calc(100vh-6.75rem)] md:max-h-[calc(100vh-3rem)]",
     unless(@live_action == :show, do: "hidden", else: "flex")
   ]}
 >
