+++ lib/bike_brigade_web/live/sms_message_live/index.html.heex
@@ -1,8 +1,8 @@
 <div class={[
-  "flex  h-full max-h-full min-h-0 -mx-3"
+  "flex flex-wrap h-full max-h-full min-h-0 -mx-3"
 ]}>
   <div class={[
-    "h-full w-full px-2 md:w-4/12 sm:block",
+    "h-full w-full px-2 sm:w-4/12 sm:block",
     unless(@live_action == :index, do: "hidden")
   ]}>
     <div class={[
@@ -51,22 +51,66 @@
           </div>
         </div>
       </div>
-      <.sms_message_list
-        rider_link_fn={fn rider_id -> ~p"/messages/#{rider_id}" end}
-        conversations={@conversations}
-      />
+      <ul
+        class="flex-grow-0 h-full max-h-full px-4 py-5 overflow-y-auto scrolling-touch overscroll-contain sm:p-6"
+        id="conversation-list"
+        phx-hook="ConversationList"
+        phx-update="append"
+      >
+        <li
+          :for={{rider, last_message} <- @conversations}
+          id={"conversation-list-item:#{rider.id}"}
+          data-rider-id={rider.id}
+        >
+          <.link
+            patch={~p"/messages/#{rider}"}
+            class="block transition duration-150 ease-in-out hover:bg-gray-200 focus:outline-none"
+          >
+            <div class="flex items-center px-2 py-2">
+              <div class="flex items-center flex-1 min-w-0">
+                <div class="flex-shrink-0 hidden mr-4 lg:block">
+                  <img class="w-12 h-12 rounded-full" src={gravatar(rider.email)} alt="" />
+                </div>
+                <div class="flex-1 min-w-0">
+                  <div class="flex flex-col lg:space-x-2 lg:items-center lg:flex-row">
+                    <div class={[
+                      "pii",
+                      "text-sm leading-5 text-indigo-600 truncate",
+                      if(last_message.from == rider.phone,
+                        do: "font-extrabold",
+                        else: "font-semibold"
+                      )
+                    ]}>
+                      <%= rider.name %>
+                    </div>
+                    <div class="text-xs font-medium">
+                      <%= datetime(last_message.sent_at) %>
+                    </div>
+                  </div>
+                  <div class={[
+                    "flex items-center mt-2 text-sm leading-5 text-gray-500",
+                    if(last_message.from == rider.phone, do: "font-extrabold", else: "font-medium")
+                  ]}>
+                    <span class="truncate">
+                      <%= preview_message(last_message) %>
+                    </span>
+                  </div>
+                </div>
+              </div>
+            </div>
+          </.link>
+        </li>
+      </ul>
     </div>
   </div>
-  <div class="flex md:w-8/12 lg:8/12 max-h-[calc(100vh-3rem)] max-h-[calc(100vh-6rem)] sm:max-h-[calc(100vh-6.75rem)] md:max-h-[calc(100vh-3rem)]">
-    <.live_component
-      :if={@selected_rider}
-      module={ConversationComponent}
-      id={@selected_rider.id}
-      rider={@selected_rider}
-      live_action={@live_action}
-      current_user={@current_user}
-    />
-  </div>
+  <.live_component
+    :if={@selected_rider}
+    module={ConversationComponent}
+    id={@selected_rider.id}
+    rider={@selected_rider}
+    live_action={@live_action}
+    current_user={@current_user}
+  />
   <div
     :if={@latest_campaign_tasks != []}
     class={"w-full px-3 sm:w-2/12 #{unless @live_action == :tasks, do: "hidden"} sm:block"}
