+++ test/support/fixtures.ex
@@ -58,21 +58,6 @@ defmodule BikeBrigade.Fixtures do
     |> Repo.preload(program: [:items])
   end
 
-  def fixture(:campaign_with_riders, attrs) do
-    campaign = fixture(:campaign)
-
-    riders = for _i <- 1..7 do
-      fixture(:rider)
-    end
-    Enum.each(riders, fn rider ->
-      Delivery.create_campaign_rider(%{
-            campaign_id: campaign.id,
-            rider_id: rider.id
-            })
-    end)
-    {campaign, riders}
-  end
-
   def fixture(:rider, attrs) do
     location = Toronto.random_location()
     {:ok, rider} =
